backend/auth/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-dotenv
sqlalchemy
pydantic
redis
pyjwt